Remove stale top-level invariants/watchdog test copies
Both files were path-duplicates of the canonical tests under executor/ and
runtime/. Before deleting, ported the assertions unique to them:
- invariants: bare {"jsonrpc": "2.0"} envelope fall-through case
- watchdog: registry.update_status(pid, "stopped") on kill, and the
  metrics_sampler-raises branch

Also converted the executor batch-verification tests from async methods on a
unittest.TestCase (which pytest-asyncio never awaited — dead no-ops) into live
module-level async functions, restoring coverage of the batch InvariantViolation
branch and removing the deprecation warnings.

Verified: 2529 passed, no per-file coverage regression, total 95.39 -> 95.40.
